chore: crypto-ffi: split the JS wrapper into smaller modules
This gets rid of the import cycle between CoreCrypto and
CoreCryptoContext. Now CoreCrypto is merely a toplevel module
that just re-exports everything from other modules and takes
care of initialising the WASM module.
